// src/types/apiDataSalesDetail.ts
export interface ApiSalesDetail extends Record<string, unknown> {
    dtl_flagmemberkhusus: string;
    dtl_namaoutlet: string;
    dtl_namasuboutlet: string;
    dtl_tglmulai: string;
    dtl_tglakhir: string;
    dtl_kodesupplier: string;
    dtl_namasupplier: string;
    index: string;
    dtl_rtype: string;
    dtl_tanggal: string;
    dtl_jam: string;
    dtl_struk: string;
    dtl_stat: string;
    dtl_kasir: string;
    dtl_no_struk: string;
    dtl_seqno: string;
    dtl_prdcd_ctn: string;
    dtl_prdcd: string;
    dtl_nama_barang: string;
    dtl_unit: string;
    dtl_frac: string;
    dtl_tag: string;
    dtl_bkp: string;
    dtl_qty_pcs: number;
    dtl_qty: number;
    dtl_harga_jual: number;
    dtl_diskon: number;
    dtl_gross: number;
    dtl_netto: number;
    dtl_hpp: number;
    dtl_margin: number;
    dtl_k_div: string;
    dtl_k_dept: string;
    dtl_k_katb: string;
    dtl_nama_div: string;
    dtl_nama_dept: string;
    dtl_nama_katb: string;
    dtl_cusno: string;
    dtl_namamember: string;
    dtl_memberkhusus: string;
    dtl_outlet: string;
    dtl_suboutlet: string;
    dtl_kategori: string;
    dtl_sub_kategori: string;
    dtl_tipemember: string;
    dtl_group_member: string;
    hgb_kodesupplier: string;
    sup_namasupplier: string;
    dtl_belanja_pertama: string;
    dtl_method: string;
    dtl_amount: number;
    dtl_virtype: string;
    dtl_cb: number;
    dtl_margin_percent: number;
}